requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.31.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msrest==0.7.1
narwhals==1.12.1
numpy
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.13
PySimpleGUI==5.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
watchfiles==0.24.0
websockets==13.1
